+++ src/api/java/com/nisovin/shopkeepers/api/ui/UIType.java
@@ -2,5 +2,8 @@ package com.nisovin.shopkeepers.api.ui;
 
 import com.nisovin.shopkeepers.api.types.Type;
 
+/**
+ * A type of user interface.
+ */
 public interface UIType extends Type {
 }