+++ src/main/java/com/nisovin/shopkeepers/ui/SKUISession.java
@@ -4,46 +4,65 @@ import org.bukkit.entity.Player;
 
 import com.nisovin.shopkeepers.api.ui.UISession;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
+import com.nisovin.shopkeepers.util.Validate;
 
-public class SKUISession implements UISession {
+public final class SKUISession implements UISession {
 
-	// Reference shopkeeper directly and not by id, because the id might change or currently be invalid
-	// (for inactive shopkeepers).. especially important for remotely opened windows.
-	private final AbstractShopkeeper shopkeeper;
 	private final UIHandler uiHandler;
 	private final Player player;
+	// Can be null for UIs that are not associated with some shopkeeper:
+	private final AbstractShopkeeper shopkeeper;
+	private boolean uiActive = true;
 	private boolean valid = true;
 
-	public SKUISession(AbstractShopkeeper shopkeeper, UIHandler handler, Player player) {
-		this.shopkeeper = shopkeeper;
-		this.uiHandler = handler;
+	public SKUISession(UIHandler uiHandler, Player player, AbstractShopkeeper shopkeeper) {
+		Validate.notNull(uiHandler, "uiHandler is null");
+		Validate.notNull(player, "player is null");
+		this.uiHandler = uiHandler;
 		this.player = player;
+		this.shopkeeper = shopkeeper; // Can be null
 	}
 
 	@Override
-	public AbstractShopkeeper getShopkeeper() {
-		return shopkeeper;
+	public final AbstractUIType getUIType() {
+		return uiHandler.getUIType();
 	}
 
-	public UIHandler getUIHandler() {
+	public final UIHandler getUIHandler() {
 		return uiHandler;
 	}
 
 	@Override
-	public AbstractUIType getUIType() {
-		return uiHandler.getUIType();
-	}
-
-	public Player getPlayer() {
+	public final Player getPlayer() {
 		return player;
 	}
 
-	void onSessionEnd() {
+	@Override
+	public final AbstractShopkeeper getShopkeeper() {
+		return shopkeeper;
+	}
+
+	@Override
+	public boolean isUIActive() {
+		return uiActive;
+	}
+
+	@Override
+	public void deactivateUI() {
+		uiActive = false;
+	}
+
+	@Override
+	public void activateUI() {
+		uiActive = true;
+	}
+
+	final void onSessionEnd() {
 		valid = false;
 	}
 
 	@Override
-	public boolean isValid() {
+	public final boolean isValid() {
 		return valid;
 	}
 }